requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
babel==2.17.0
banks==2.1.3
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
chromedriver-autoinstaller==0.6.4
click==8.2.1
colorama==0.4.6
courlan==1.3.2
cssselect==1.3.0
dataclasses-json==0.6.7
dateparser==1.2.2
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
fastapi==0.115.12
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
filetype==1.2.0
firecrawl-py==2.12.0
frozenlist==1.7.0
fsspec==2025.5.1
greenlet==3.2.3
griffe==1.7.3
h11==0.16.0
html2text==2024.2.26
htmldate==1.9.3
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jusText==3.0.2
langchain==0.3.25
langchain-core==0.3.64
langchain-text-splitters==0.3.8
langsmith==0.3.45
llama-cloud==0.1.26
llama-cloud-services==0.6.34
llama-index==0.12.44
llama-index-agent-openai==0.4.12
llama-index-cli==0.4.3
llama-index-core==0.12.44
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.7.7
llama-index-instrumentation==0.2.0
llama-index-llms-openai==0.4.7
llama-index-multi-modal-llms-openai==0.5.1
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.9
llama-index-readers-llama-parse==0.4.0
llama-index-readers-web==0.4.2
llama-index-workflows==1.0.1
llama-parse==0.6.34
lxml==5.4.0
lxml_html_clean==0.4.2
mangum==0.19.0
markdownify==1.1.0
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.6.2
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
newspaper3k==0.2.8
ngrok==1.4.0
nltk==3.9.1
numpy==2.2.6
openai==1.84.0
orjson==3.10.18
outcome==1.3.0.post0
oxylabs==2.0.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pinecone==7.0.2
pinecone-plugin-assistant==1.6.1
pinecone-plugin-interface==0.0.7
platformdirs==4.3.8
playwright==1.53.0
propcache==0.3.2
pydantic==2.11.5
pydantic_core==2.33.2
pyee==13.0.0
pypdf==5.7.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
readability==0.3.2
readability-lxml==0.8.4.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
selenium==4.34.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spider-client==0.0.27
SQLAlchemy==2.0.41
starlette==0.46.2
striprtf==0.0.26
tenacity==9.1.2
tiktoken==0.9.0
tinysegmenter==0.3
tld==0.13.1
tldextract==5.3.0
tqdm==4.67.1
trafilatura==2.0.0
trio==0.30.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.3
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.1
zstandard==0.23.0
